chore: separate AttendanceSession
